LIBHYDRA-317. Sends resource update to Fedora
Updated "app/controllers/resource_controller.rb" to actually send
a SPARQL query describing the update to Fedora via the Plaston "update"
command.

The message is sent synchronously, and redirects back to the
metadata detail page if the update was successful. If the update did
not succeed, the edit form is shown again.

https://issues.umd.edu/browse/LIBHYDRA-317