LIBHYDRA-317. Attempt at providing updated data on failure
THIS CODE SHOULD _NOT_ BE CHECKED IN WITHOUT REVISIONS

This code is an attempt to regenerate the updated data when a
Plastron submission fails. It is set to intentionally fail by
appending "_FORCE_FAILURE" onto the id submitted to Plastron.

This was intended to demonstrate issues with the "diff" method
in the presense of failures -- it is possible to show the
updated data after the failure, then when the form is submitted
the changes from the previous update are lost.

https://issues.umd.edu/browse/LIBHYDRA-317