DELAYED_JOB. DownloadUrlExpirationJob with self-scheduling
A DownloadUrlExpirationJob class that enqueues the next run for the
job, and then runs the job.

This behavior is probably not very useful, because there is no way to
detect if a job has already been scheduled from another run of the
application, which means there may be multiple instances of the
job in the schedules queue.