LIBFCREPO-1471. Added "dotenv" gem and ".env" files
Added the "dotenv" gem to the "Gemfile" (the "dotenv-rails" gem is no
longer needed -- we can use "dotenv" directly).

Added "env_example" and ".env.development" files from the legacy
Archelon.

Modified the ".gitignore" file as in legacy Archelon to enable the
".env.development" file to be included in Git.

Rails 7.1 uses the “foreman” gem to run the asset pipeline watchers and
the Rails server, via the “bin/dev” script. As described in
<https://github.com/ddollar/foreman/issues/702>, “foreman” and the
“dotenv” gem conflict when reading in the environment variables (any
environment variables in the CLI environment are ignored, instead of
overriding the variables in the various “.env” files).

Used the workaround of adding `--env=/dev/null` to the “foreman” command
in the “bin/dev” script so that the CLI environment variables and
“.env” files will work as expected (i.e., that the CLI environment
variables take precedence over the “.env” files).

https://umd-dit.atlassian.net/browse/LIBFCREPO-1471