LIBFCREPO-1471. Removed useless assignment in "from_subquery"
Ruby LSP pointed out that the first "document_hash" assignment in
the "from_subquery" method was useless, because of the assignment on
the next line.

Changed to simply returning the value of
`args[:document][subquery_field]['docs']`
because that was the simplest way to express what is happening.

Note that in legacy Archelon, the code for this method was

```
  def from_subquery(subquery_field, args)
    args[:document][args[:field]] = args[:document][subquery_field]['docs']
  end
```

but the reassignment to `args[:document][args[:field]]` is no longer
possible, and doesn't seem to have any effect in any case.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1471