LIBFCREPO-1471. Added React component loading to "application.js"
Updated the "application.js" file to load the React JSX components
from the "app/javascript/components" directory (or subdirectories).

This relies on the “esbuild-plugin-import-glob” plugin to allow use of
`import components from "./components/**/*.jsx"` to find all the React
JSX components in the “app/javascript/components/” directory
(and subdirectories) and then uses the “ReachRailsUJS” (from the
“react-rails” gem) so that the components are usable in ERB templates
using `<%= react_component(...) %>` syntax.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1471