LIBFCREPO-1471. Updated configuration to use Postgres in production
Updated the "config/database.yml" file to use Postgres database for
"production".

The "development" and "test" environments still use SQLite, but the
SQLite database is now stored in the Rails 7 default of "storage/"
directory, instead of "db/".

Added the "pg" gem to the "Gemfile" and ran:

```
$ bundle install
```

Modified the ".env.development" file to have the SQLite file stored in
the Rails 7 default "storage/" directory instead of "db/".

https://umd-dit.atlassian.net/browse/LIBFCREPO-1471