LIBFCREPO-1471. Added UMD logo and application title to navbar
Copied the "app/assets/images/blacklight/logo.png" file from legacy
Archelon, which contains the UMD logo image.

Copied the "top_navbar_component.html.erb" file from Blacklight 8, and
modified to support the UMD logo and application title.

Added "_variables.scss" file to override SCSS variables in other SCSS
files. Needed to do this for the logo because Propshaft does not
recognize the "image_url" helper that Sprockets does, so the reference
needed to be change to "url". See <https://stackoverflow.com/a/76429854>.

Added a "custom.css" file to be consistent with legacy Archelon, and
added in CSS customizations for the logo and application title.

Updated the "application.bootstrap.scss" file to import the
"_variables.scss", and "custom.scss" files

https://umd-dit.atlassian.net/browse/LIBFCREPO-1471