LIBFCREPO-1471. Changed Action Cable subscription adapter  to “async”
Changed Action Cable subscription adapter in “config/cable.yml” to “async”
(from stock “redis”), to match the configuration in legacy Archelon.

This removes the need to add the “redis” gem to support Action Cable.

The Action Cable functionality, used to update the job statuses, is
so lightly used, that continuing to use the "async" adapter seemed
reasonable.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1471