LIBFCREPO-1471. Added "Manage" drop-down to navigation bar
The base “Dockerfile” provided by Rails 7.1 sets a “rails” user as the
container’s default user (instead of the “root” user used in legacy
Archelon). This was making the “/var/opt/archelon” directory used for
imports unusable, because this directory is modified as part of the
“k8s-fcrepo” configuration to be writable only the the user with a
uid of “2200” (which is the uid of “plastron” user in
"Dockerfile.sftp").

Modified the “Dockerfile” to set the “uid” and “gid” of the “rails” user
to 2200, so that it could write to the “/var/opt/archelon” directory
and its subdirectories, which includes the “import” directory).

https://umd-dit.atlassian.net/browse/LIBFCREPO-1471